+++ packages/react-core/src/components/DescriptionList/examples/DescriptionListCompact.tsx
@@ -5,7 +5,7 @@ import {
   DescriptionListGroup,
   DescriptionListDescription
 } from '@patternfly/react-core';
-import PlusCircleIcon from '@patternfly/react-icons/dist/esm/icons/plus-circle-icon';
+import RhUiAddCircleFillIcon from '@patternfly/react-icons/dist/esm/icons/rh-ui-add-circle-fill-icon';
 
 export const DescriptionListCompact: React.FunctionComponent = () => (
   <DescriptionList isCompact aria-label="Compact example">
@@ -26,7 +26,7 @@ export const DescriptionListCompact: React.FunctionComponent = () => (
     <DescriptionListGroup>
       <DescriptionListTerm>Pod selector</DescriptionListTerm>
       <DescriptionListDescription>
-        <Button variant="link" isInline icon={<PlusCircleIcon />}>
+        <Button variant="link" isInline icon={<RhUiAddCircleFillIcon />}>
           app=MyApp
         </Button>
       </DescriptionListDescription>